make sure is int

+++ config/tasks.php
@@ -1,7 +1,7 @@
 <?php
 
 return [
-    'default_max_attempts' => env('MAX_ATTEMPTS', 3),
-    'default_lease_seconds' => env("LEASE_SECONDS", 60),
-    'default_timeout_seconds' => env('TASK_TIMEOUT_SECONDS', 300),
+    'default_max_attempts' => (int) env('MAX_ATTEMPTS', 3),
+    'default_lease_seconds' => (int) env("LEASE_SECONDS", 60),
+    'default_timeout_seconds' => (int) env('TASK_TIMEOUT_SECONDS', 300),
 ];
